moved all formatting funcs from init to a separate file
